+++ CHANGELOG.md
@@ -6,11 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+* Support SSE-C encryption mode for files
 
 ## [0.5.0] - 2021-05-31
 
 ### Added
-* Support SSE-C encryption mode for files
 * Support isFileLockEnabled for buckets
 * Support defaultRetention for buckets